% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdmm_amp}
\alias{cdmm_amp}
\title{cdmm_amp}
\usage{
cdmm_amp(y, x, fac, lams, mu, maxv, maxit, tol)
}
\arguments{
\item{y}{Outcome vector}

\item{x}{OTU matrix}

\item{fac}{Vector of scaling factors for OTU matrix}

\item{lams}{Sequence of lambda values to try}

\item{mu}{Lagrange scale parameter}

\item{maxv}{Max number of nonzero variables}

\item{maxit}{Max number of iterations allowed}

\item{tol}{Parameter defining convergence}
}
\description{
Runs Lin's compositional lasso
}
