% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_cdmm.R
\name{cv_cdmm}
\alias{cv_cdmm}
\title{cv_cdmm}
\usage{
cv_cdmm(y, x, nfolds, lamseq, nlam = 100, rlam = 1/nlam, mu = 1,
  std = TRUE, maxv = 0.4 * length(y), maxit = c(20, 50), tol = c(1e-04,
  1e-07))
}
\arguments{
\item{y}{Vector of outcomes}

\item{x}{Matrix of log relative abundances}

\item{nfolds}{Number of folds for cross validation}

\item{lamseq}{Sequence of lambda values to try}

\item{nlam}{Number of lambda values to try; default 100}

\item{rlam}{Minimum lambda value as a fraction of maxlam; default 1/nlam}

\item{mu}{Lagrange scale parameter; default 1}

\item{std}{Logical indicating whether data should be standardized; default TRUE}

\item{maxv}{Maximum number of nonzero variables; default = 0.4*n}

\item{maxit}{Maximum number of iterations (two measures of convergence); default = c(20, 50)}

\item{tol}{Tolerance for convergence (same two measures); default = c(1e-4, 1e-7)}
}
\description{
Chooses lambda for the compositional lasso by cross-validated prediction error
}
