% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_cl.R
\name{gic.cdmm}
\alias{gic.cdmm}
\title{gic.cdmm}
\usage{
gic.cdmm(y, x, nlam)
}
\arguments{
\item{y}{Vector of outcomes}

\item{x}{Matrix of log relative abundances}

\item{nlam}{Number of lambda values to try; default 100}
}
\description{
Wrapper function for C++ function that runs the compositional lasso,
using the generalized information criterion (GIC) to select lambda
}
